Don't crash on 0 pointer.
svn path=/trunk/kdepim/; revision=366752

+++ kresources/opengroupware/kabc_resourceopengroupware.cpp
@@ -134,7 +134,7 @@ void ResourceOpenGroupware::doClose()
 {
   kdDebug() << "ResourceOpenGroupware::doClose()" << endl;
 
-  mDownloadJob->kill();
+  if ( mDownloadJob ) mDownloadJob->kill();
 }
 
 bool ResourceOpenGroupware::load()
